@@ -71,74 +71,69 @@
     ]
   },
   "dependencies": {
-    "@angular/cdk-experimental": "^7.2.0",
-    "@nrwl/nx": "7.3.0",
-    "@types/fontfaceobserver": "^0.0.6",
-    "fontfaceobserver": "^2.1.0",
-    "resize-observer-polyfill": "^1.5.1",
-    "@ngrx/effects": "6.1.2",
-    "@ngrx/router-store": "6.1.2",
-    "@ngrx/store": "6.1.2"
+    "@nrwl/nx": "7.3.0"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "0.11.4",
     "@angular-devkit/build-ng-packagr": "0.11.4",
-    "@angular-devkit/core": "7.1.4",
-    "@angular-devkit/schematics": "7.1.4",
-    "@angular/animations": "7.1.4",
-    "@angular/cdk": "7.2.0",
-    "@angular/cli": "7.1.4",
-    "@angular/common": "7.1.4",
-    "@angular/compiler": "7.1.4",
-    "@angular/compiler-cli": "7.1.4",
-    "@angular/core": "7.1.4",
+    "@angular-devkit/core": "7.2.0-rc.0",
+    "@angular-devkit/schematics": "7.2.0-rc.0",
+    "@angular/animations": "7.2.0",
+    "@angular/cdk": "7.2.1",
+    "@angular/cli": "7.2.0-rc.0",
+    "@angular/common": "7.2.0",
+    "@angular/compiler-cli": "7.2.0",
+    "@angular/compiler": "7.2.0",
+    "@angular/core": "7.2.0",
     "@angular/flex-layout": "7.0.0-beta.22",
-    "@angular/forms": "7.1.4",
-    "@angular/language-service": "7.1.4",
-    "@angular/material": "7.2.0",
-    "@angular/platform-browser": "7.1.4",
-    "@angular/platform-browser-dynamic": "7.1.4",
-    "@angular/router": "7.1.4",
+    "@angular/forms": "7.2.0",
+    "@angular/language-service": "7.2.0",
+    "@angular/material": "7.2.1",
+    "@angular/platform-browser-dynamic": "7.2.0",
+    "@angular/platform-browser": "7.2.0",
+    "@angular/router": "7.2.0",
     "@graphql-modules/logger": "^0.2.3",
-    "@nrwl/angular-console-enterprise-electron": "0.0.35",
-    "@nrwl/angular-console-enterprise-frontend": "0.0.35",
+    "@nrwl/angular-console-enterprise-electron": "0.0.38",
+    "@nrwl/angular-console-enterprise-frontend": "0.0.38",
     "@nrwl/builders": "7.3.0",
     "@nrwl/schematics": "7.3.0",
     "@types/electron-store": "^1.3.0",
     "@types/express": "4.16.0",
+    "@types/fontfaceobserver": "^0.0.6",
     "@types/get-port": "^4.0.1",
     "@types/graphql": "14.0.3",
     "@types/jest": "^23.3.10",
     "@types/node": "8.10.38",
+    "@types/opn": "^5.1.0",
     "@types/request": "^2.48.1",
     "@types/semver": "^5.5.0",
     "@types/shelljs": "^0.8.1",
-    "@types/sinon": "5.0.7",
-    "@types/sinon-chai": "3.2.1",
+    "@types/sinon-chai": "3.2.2",
+    "@types/sinon": "7.0.3",
     "@types/strip-json-comments": "^0.0.30",
     "@types/tmp": "^0.0.33",
     "@types/universal-analytics": "0.4.2",
-    "apollo-angular": "1.5.0",
     "apollo-angular-link-http": "1.4.0",
-    "apollo-cache-inmemory": "1.3.11",
+    "apollo-angular": "1.5.0",
+    "apollo-cache-inmemory": "1.3.12",
+    "apollo-link-error": "1.1.5",
     "apollo-link": "1.2.4",
-    "apollo-link-error": "1.1.2",
     "apollo-server-express": "^2.2.2",
     "codelyzer": "4.5.0",
     "core-js": "2.5.7",
     "cypress": "3.1.1",
     "dotenv": "6.2.0",
-    "electron": "2.0.14",
     "electron-builder": "20.28.4",
     "electron-installer-dmg": "2.0.0",
     "electron-packager": "13.0.1",
     "electron-rebuild": "^1.8.2",
     "electron-store": "2.0.0",
     "electron-updater": "3.1.2",
+    "electron": "2.0.14",
     "express": "4.16.4",
     "fix-path": "2.1.0",
+    "fontfaceobserver": "^2.1.0",
     "get-port": "4.0.0",
-    "graphql": "^0.13.1",
     "graphql-code-generator": "0.15.2",
     "graphql-codegen-typescript-apollo-angular": "0.15.2",
     "graphql-codegen-typescript-client": "0.15.2",
